+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Fixed
 
 - 🐛(ci) fix the CHANGELOG check failing when a pull request is opened
+- 🐛(back) replace whitenoise with servestatic for async-capable serving
 
 ## [0.0.24] - 2026-09-08